Extract test schemas
