+++ test/surface/components/form/inputs_test.exs
@@ -5,29 +5,7 @@ defmodule Surface.Components.Form.InputsTest do
   alias Surface.Components.Form.Field
   alias Surface.Components.Form.Inputs
   alias Surface.Components.Form.TextInput
-
-  defmodule Parent do
-    defmodule Child do
-      use Ecto.Schema
-
-      embedded_schema do
-        field(:name, :string)
-      end
-
-      def changeset(cs_or_map, data), do: Ecto.Changeset.cast(cs_or_map, data, [:name])
-    end
-
-    use Ecto.Schema
-
-    embedded_schema do
-      embeds_many(:children, Child)
-    end
-
-    def changeset(cs_or_map \\ %__MODULE__{}, data),
-      do:
-        Ecto.Changeset.cast(cs_or_map, data, [])
-        |> Ecto.Changeset.cast_embed(:children)
-  end
+  alias Surface.Schemas.Parent
 
   test "using generated form received as slot arg" do
     html =
